tuwitter kadai

+++ app/Http/Controllers/TasksController.php
@@ -4,21 +4,32 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 
+use Illuminate\Support\Facades\Auth;
+
 use App\Task;
 
 class TasksController extends Controller
 {
   
-    // getでmessages/にアクセスされた場合の「一覧表示処理」
+    // getでtask/にアクセスされた場合の「一覧表示処理」
     public function index()
     {
-        //タスク一覧取得
-        $tasks = Task::all();
-        
-        //タスク一覧ビューでそれを表示
-        return view('tasks.index', [
-            'tasks' => $tasks,
-            ]);
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(15);
+
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            return view('tasks.index', $data);
+        }
+
+        // Welcomeビューでそれらを表示
+        return view('welcome', $data);
     }
 
     // getでtasks/createにアクセスされた場合の「新規登録画面表示処理」
@@ -37,11 +48,13 @@ class TasksController extends Controller
         
         //バリテーション
         $request->validate([
+            //'user_id' => 'required|max:255',   // 追加
             'content' => 'required|max:255',
             'status' => 'required|max:10',
             ]);
          // タスクを作成
         $task = new Task;
+        $task->user_id = Auth::user()->id;   // 追加
         $task->content = $request->content;
         $task->status = $request->status;
         $task->save();
@@ -80,12 +93,14 @@ class TasksController extends Controller
     public function update(Request $request, $id)
     {
         $request->validate([
+         //   'user_id' => 'required|max:255',   // 追加
             'content' => 'required|max:255',
             'status' => 'required|max:10',
             ]);
          // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
         // タスクを更新
+       // $task->user_id = $request->user_id;
         $task->content = $request->content;
         $task->status = $request->status;
         $task->save();
@@ -101,7 +116,10 @@ class TasksController extends Controller
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
         // メッセージを削除
-        $task->delete();
+         // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
 
         // トップページへリダイレクトさせる
         return redirect('/');
